backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
numpy==1.26.4
scipy==1.13.0
rasterio==1.3.10
scikit-learn==1.4.2
xgboost==2.0.3
pandas==2.2.2
shapely==2.0.4
pyproj==3.6.1
Pillow==10.3.0
python-multipart==0.0.9
aiofiles==23.2.1
